New name for @cd first\ node\ app/. ParametersTest

+++ 06.DataTypes/06.DataTypes/AppDelegate.m
@@ -120,7 +120,7 @@
     return YES;
 }
 
-- (int)test:(NSInteger)someVar testPointer:(NSInteger*)testPointer{
+- (long)test:(NSInteger)someVar testPointer:(NSInteger*)testPointer{
     *testPointer = 5;
     return someVar;
 }
